chore: unify examples and templates into a single source
